SignUp component created
