+++ src/Components/App/index.js
@@ -6,6 +6,7 @@ import { useState } from "react";
 
 function App() {
   const [welcome, setWelcome] = useState(false);
+  const [name, setName] = useState("Username");
 
   if (welcome)
     return (
@@ -15,9 +16,11 @@ function App() {
       </div>
     );
 
+  // useEffect that fetched data using the 'name' to send a GET request for the correct table
+
   return (
     <div className="main-app">
-      <WelcomePage setWelcome={setWelcome} />
+      <WelcomePage setWelcome={setWelcome} setName={setName} />
       <p>
         Input and display will be hidden until welcome page has been passed.
       </p>
