+++ src/Components/WelcomePage/index.js
@@ -1,35 +1,13 @@
 import React from "react";
-import { BsArrowRightSquare } from "react-icons/bs";
-import { Box, TextField } from "@mui/material";
+import SignUp from "../SignUp";
 import "./WelcomePage.css";
 
-function WelcomePage({ setWelcome }) {
-  function handleClick(e) {
-    console.log("ONCLICK FUNCTION");
-    setWelcome(true);
-  }
+function WelcomePage({ setWelcome, setName }) {
+  // onchange function that takes the users name and passes back up to app
 
   return (
     <div>
-      <h1>Hey there, Camper!</h1>
-      <div id="welcome-input">
-        <Box
-          component="form"
-          sx={{
-            "& > :not(style)": { m: 1, width: "25ch" },
-          }}
-          noValidate
-          autoComplete="off"
-        >
-          <TextField label="What shall I call you?" variant="standard" />
-        </Box>
-        <BsArrowRightSquare
-          id="enter-icon"
-          onClick={(e) => {
-            handleClick(e);
-          }}
-        />
-      </div>
+      <SignUp setWelcome={setWelcome} setName={setName} />
     </div>
   );
 }
